+++ test/vlog/pp8.v
@@ -1,0 +1,10 @@
+`define add(x, y) ((x) + (y))
+`add(1, 2)
+hello
+`define ASSERT assert
+`ASSERT(x == y);
+`add((1 + 2), ident)
+`add(1) // Error
+`baz(4, 5, 6) // Warning
+`define foo(fff)
+`foo(4, 5, 6) // Error